Adding AI that find minimum variance for KSrho input for CH4 molecule

+++ tests/tEFF/AI_angdist_sim.py
@@ -8,7 +8,7 @@ import time
 sys.path.append("../../")
 from pyBall import eFF as eff
 elementPath = "export/scan_data/angdistscan_CH4.xyz"
-fileToSavePath = "results/AI/result_1.txt"
+fileToSavePath = "results/AI/result_3.txt"
 
 def count_mask_lines( fgo_file, mask='#iconf' ):
     line = None
@@ -57,8 +57,33 @@ def extract_blocks(xyz_file):
             params[key][i] = val
     return params, nrec
 
+def plot_energy_landscape( Xs, Ys, Es, Espan=None):
+    """Plot energy landscape from XYZ file using imshow (simple and robust)."""
+    #params, nrec = extract_blocks(xyz_file)
+    xs, idx = np.unique ( Xs, return_inverse=True )
+    ys, idy  = np.unique( Ys, return_inverse=True )
+    energy_grid = np.full((len(xs), len(ys)), np.nan)
+    # Fill the grid
+    for i in range(len(Xs)):
+        energy_grid[ idx[i], idy[i]] = Es[i]
+    # Plot
+    plt.figure(figsize=(10,8))
+    if Espan is not None: 
+        vmin = np.nanmin(energy_grid)
+        vmax = vmin + Espan
+    else:
+        vmin = None
+        vmax = None
+    plt.imshow(energy_grid, extent=[ys.min(), ys.max(), xs.max(), xs.min()], aspect='auto', cmap='inferno', vmin=vmin, vmax=vmax)
+    plt.colorbar(label='Total Energy ')
+    plt.xlabel('Distance (Å)')
+    plt.ylabel('Angle (rad)')
+    #plt.title('Potential Energy Surface')     
+
+
 def save_simulation(angleArr, distArr, allEtot, flexVar, variance, fileToSavePath):
-    os.remove(fileToSavePath)
+    if os.path.exists(fileToSavePath):
+        os.remove(fileToSavePath)    
     with open(fileToSavePath, "w") as f:
         # Convert and flatten arrays (to ensure 1D), then write manually
         angleArrnp = np.array(angleArr).flatten()
@@ -68,19 +93,12 @@ def save_simulation(angleArr, distArr, allEtot, flexVar, variance, fileToSavePat
         f.write(" ".join(f"{val:.6f}" for val in angleArrnp) + "\n")
         f.write(" ".join(f"{val:.6f}" for val in distArrnp) + "\n")
         f.write("\n\n\n")  # 3 blank lines
-        print(flexVar)
 
         for i in range(len(flexVar)):
             # Ensure each input is 1D
-            print(flexVar)
-
             flex_row = np.array(flexVar[i]).flatten()
-            print(flex_row)
             var_row = np.array(variance[i]).flatten()
-            print(var_row)
             etot_row = np.array(allEtot[i]).flatten()
-            print(etot_row)
-
 
             # Write each on a single line
             f.write(" ".join(f"{val:.6f}" for val in flex_row) + "\n")
@@ -88,11 +106,22 @@ def save_simulation(angleArr, distArr, allEtot, flexVar, variance, fileToSavePat
             f.write(" ".join(f"{val:.6f}" for val in etot_row) + "\n")
             f.write("\n")  # blank line between each block
 
+def get_variance(theta):
+    outEs = np.zeros((nrec,5))
+    eff.processXYZ(elementPath, bOutXYZ=True, outEs=outEs, bCoreElectrons=bCoreElectrons, bChangeCore=False, bChangeEsize=True, nstepMax=10000, dt=0.005, Fconv=1e-3, ialg=2, KRSrho=theta )
+    outEsdiff = np.zeros((nrec,5))
+    outEsdiff[:,0] = params["Etot"] - outEs[:,0]
+    outEsdiff[:,0] -= sum(outEsdiff[:,0])/len(outEsdiff[:,0])
+    outEsdiff[:,0][params['dist'] == 0.7] = 0 #getting rid of unwanted valuess 
+    var = np.var(outEsdiff[:,0])
+
+    return var, outEsdiff[:,0]
+
+
 if __name__ == "__main__":
     print("#=========== RUN /home/gabriel/git/FireCore/tests/tEFF/AI_rough.py")
 
-    eff.setVerbosity(1,0)
-    print("verbos")
+    eff.setVerbosity(0,0)
     atomParams = np.array([
     #  Q   sQ   sP   cP
     [ 0.,  1.0, 1.00, 0.0 ], # 0
@@ -119,47 +148,97 @@ if __name__ == "__main__":
 
     angleArr = params['ang']
     distArr = params['dist']
-    # using lists for better performance
+    
     allEtot = []
     flexVar = []
     variance = []
+    iteretions = 50
+    theta0 = np.array([1.125, 0.9, -0.2])
+    # theta0 = np.array([1.1, 0.7, -0.4])
 
-    for i in range(2):
-        outEs = np.zeros((nrec,5))
-        eff.processXYZ(elementPath, bOutXYZ=True, outEs=outEs, bCoreElectrons=bCoreElectrons, bChangeCore=False, bChangeEsize=True, nstepMax=10000, dt=0.005, Fconv=1e-3, ialg=2 );
-        outEsdiff = np.zeros((nrec,5))
-        outEsdiff[:,0] = params["Etot"] - outEs[:,0]
-        outEsdiff[:,0] -= sum(outEsdiff[:,0])/len(outEsdiff[:,0])
-        outEsdiff[:,0][params['dist'] == 0.7] = 0 #getting rid of unwanted valuess 
-        var = np.var(outEsdiff[:,0])
+    steps = np.array([0.1, 0.1, 0.1])
+    # steps = np.array([0.05, 0.05, 0.05])    
+    # steps = np.array([0.01, 0.01, 0.01])
 
-        allEtot.append(outEsdiff[:,0])
-        theta = np.array([i, i+1])
-        print(theta)
-        flexVar.append(theta)
-        variance.append(var)
+    precision = 0.001
+    deltaTheta = np.array([0.0,0.0,0.0])
+
+    normalConst = np.sqrt(len(theta0))/len(theta0) #such that each vector of movement is on ELIPSE that enclose each point, calculated through the step method
+
+    for j in range(iteretions):
+        var0, outEsTot0 = get_variance(theta0)
+        
+        for i, step in enumerate(steps): 
+            # if i != 0:
+            #     break
+            var, outEsTot = var0.copy(), outEsTot0.copy()
+            theta1 = theta0.copy()
+            theta1[i] = theta1[i] + step
+            theta2 = theta0.copy()
+            theta2[i] = theta2[i] - step
+            
+            var1, outEsTot1 = get_variance(theta1)
+            var2, outEsTot2 = get_variance(theta2)
+            
+            
+            newStep = step
+            if var1 < var2:
+                var = var1
+                outEsTot = outEsTot1
+                theta = theta1.copy()
+                deltaTheta[i] = step
+
+                if var1 > var0:
+                    newStep = step/2
+            else:
+                var = var2
+                outEsTot = outEsTot2
+                theta = theta2.copy()
+                deltaTheta[i] = -step
+                
+                if var2 > var0:
+                    newStep = step/2
+                
+            steps[i] = newStep
+
+            allEtot.append(outEsTot.copy())
+            flexVar.append(theta.copy())
+            variance.append(var)
+            print("----------------------------------")
+            print("Theta1: ", theta1)
+            print("Theta0: ", theta0)
+            print("Theta2: ", theta2)
+            print("Var1: ", var1)
+            print("Var0: ", var0)
+            print("Var2: ", var2)
+            print("Theta: ", theta)
+            print("Iteretion: ", j)
+            print("Value index: ", i)
+            print("Step: ", newStep)
+            print("Variance: ", var)
+            # print("--------------------")
+            # input("Press Enter to continue...")
+
+        print(deltaTheta)
+        print(theta0)
+        theta0 = theta0 + deltaTheta.copy()*normalConst
         print("======================================================================================================")
-        print(i)
+        print("Theta: ", theta)
+        print("Theta0: ", theta0)
+        print("Steps: ", steps)
+        print("Variance: ", var)
         print("======================================================================================================")
 
+        if ((steps < precision) & (steps > -precision)).all():
+            print("PRECISSION MET")
+            break
 
     save_simulation(angleArr, distArr, allEtot, flexVar, variance, fileToSavePath)
 
-    print(angleArr)
-    print(distArr)
-    print(allEtot[0])
-    print(len(allEtot[0]))
-    print(len(angleArr))
-    print(len(distArr))
-    
-
-    plot_energy_landscape( angleArr, distArr, allEtot[0])
-    plt.title("Initial for theta = {flexVar[0]} and variance = {variance[0]}")
-
-    # plot_energy_landscape( params['ang'], params['dist'], outEsdiff[:,0] )
-
-    # plt.title("Difference between DFT and eFF")
-    # plt.savefig("map2d_diff.png")
+    print("======================================================================================================")
+    print("FlexVar final: ", flexVar)
+    print("Theta final: ", theta)
+    print("Steps final: ", steps)
+    print("Variance final: ", var)
 
     print("#=========== DONE /home/gabriel/git/FireCore/tests/tEFF/AI_angdist_sim.py")
-    plt.show()
